:recycle: 메시지 관련 엔티티 코드 이동
